refactor(webapp_v2): drop SourcedInput A/B scaffolding, keep glued-siblings only
The glued-siblings variant won the visual review, so the A/B framework
gets cleaned up:

- Delete SingleOutline.jsx + variantContext.js + GluedSiblings.jsx
  (GluedSiblings logic is now inlined into index.jsx as the sole
  implementation).
- Remove the preview Switch + SourcedInputVariantProvider wrapping from
  CredentialsTab.jsx; the tab body is back to its original shape.
- Trim SourcedInput.module.css to just .pickerSibling + .inputSibling
  (single-outline classes were the bulk of the file).
- Drop SOURCE_ICONS and the lucide imports it pulled into
  secretsCodec.js — the picker is text-only.
- Update COMPONENTS.md to describe the final picker shape without the
  A/B note.

Net diff: -421 lines of scaffolding, +163 lines of the kept design.
