+++ webapp_v2/COMPONENTS.md
@@ -226,7 +226,7 @@ import TextInput from '@/components/TextInput'
 ```
 
 ### `SourcedInput`
-Input paired with an optional credential source picker (Manual / Vault KV / AWS Secrets Manager / AWS IAM Role). The picker is a chip rendered inside the input's outline, showing the label of the selected source. When `sources` is empty/single, only the input renders.
+Input paired with an optional credential source picker (Manual / Vault KV / AWS Secrets Manager / AWS IAM Role) glued to its left. The picker carries the seam border so the two components meet at a single shared edge and read as one control. When `sources` is empty or has a single entry, only the input renders.
 ```jsx
 import SourcedInput from '@/components/SourcedInput'
 
@@ -244,7 +244,7 @@ import SourcedInput from '@/components/SourcedInput'
 // Sizes match Mantine inputs — default `sm`, accepts xs/sm/md/lg/xl:
 <SourcedInput size="md" {...props} />
 ```
-Supports `type="text" | "password" | "textarea"` and `size="xs" | "sm" | "md" | "lg" | "xl"` (default `sm`, matching Mantine's input default). Heights track Mantine's `--input-height-*` variables so a `size="md"` SourcedInput lines up with a `size="md"` TextInput on the same form. Renders descriptions through `MarkdownText` so inline links work. The picker layout is currently under A/B evaluation (`SourcedInputVariantContext` in `components/SourcedInput/variantContext.js`) — once the design is locked the context goes away and the chosen variant becomes the default. Textarea inputs always render the picker stacked above instead of inline (multi-line + horizontal picker looks broken).
+Supports `type="text" | "password" | "textarea"` and `size="xs" | "sm" | "md" | "lg" | "xl"` (default `sm`, Mantine's input default). Heights track Mantine's `--input-height-*` variables so a `size="md"` SourcedInput lines up with a `size="md"` TextInput on the same form. Renders the field description through `MarkdownText` so inline `[label](url)` links work. Textareas render the picker stacked above instead of inline — multi-line + horizontal picker doesn't read cleanly.
 
 ### `PasswordInput`
 Password / secret input with visibility toggle.
